% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_auto_crop.R
\name{colony_auto_crop}
\alias{colony_auto_crop}
\title{Cropping the phenotypic image of the colony array on agar plates into single patches.}
\usage{
colony_auto_crop(expr_mode = "swarming", img_gray, img_ori,
  smth_win = 0.09, row_num, col_num)
}
\arguments{
\item{expr_mode}{a string indicates the phenotypic experiment type, as different experiments require differnt
processing steps}

\item{img_gray}{an 'Image' object or an array stores the raw image in gray mode and specific size}

\item{img_ori}{an 'Image' object or an array stores the raw phenotypic image in color or gray mode}

\item{smth_win}{the length of the Gaussian smoothing window of the raw or column sum of the image intensity,if an integer,represents
number of items, else, if a value between 0 and 1, represents the proportion of the input vector}

\item{row_num}{an integer,represents the number of row within the raw image}

\item{col_num}{an integer,represents the number of row within the raw image}
}
\value{
a list contains the segmented colony patches,both in color and gray mode, and R,G,B channels(for the
        phenotypes rely on color change),as well as the row and column intensity sum and cutting coordinates
}
\description{
Before the processing and extraction of phenotypic signitures,if the experiment was done on agar plates,
it is recommeded to crop the raw images into patches containing single colony using \code{colony_auto_crop}
}
\examples{
img_patch<-single_colony_seg(expr_mode="swarming",img_gray=img$gray_target,img_ori=img$ori_img,smth_win=0.09,
                             row_num=7,col_num=7)
}
\seealso{
\code{\link{colony_manual_crop}}
}
