% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_manual_crop.R
\name{colony_manual_crop}
\alias{colony_manual_crop}
\title{Cropping of the phenotypic image of the colony(colonies) on single culture dish into single patches.}
\usage{
colony_manual_crop(expr_mode = "swarming", img_gray, img_ori,
  row_num = 5, col_num = 1, cut_coord)
}
\arguments{
\item{expr_mode}{a string indicates the phenotypic experiment type, as different experiments require differnt
processing steps}

\item{img_gray}{an 'Image' object or an array stores the raw image in gray mode and specific size}

\item{img_ori}{an 'Image' object or an array stores the raw phenotypic image in color or gray mode}

\item{row_num}{an integer,represents the number of coloies within the culture dish}

\item{col_num}{an integer,represents the number of row within the raw image(for the condition of only a few colonies,fix it to 1)}

\item{cut_coord}{a matrix with the dimension row_num*(col_num*4)(4 is the number of combined position of x and y direction,which
could be obtained based on the interactive visualization using EBImage)}
}
\value{
a list contains the segmented colony patches,both in color and gray mode, and R,G,B channels(for the
        phenotypes rely on color change),as well as the dimension of the preliminary cropped rectangular image
}
\description{
Before the processing and extraction of phenotypic signitures ,if the experiment was in quite small scale,
it is recommeded to get the cropped patch(es) containing single colony based on manually decided coordinates
using \code{colony_manual_crop}
}
\examples{
img_patch<-single_colony_seg(expr_mode="swarming",img_gray=img$gray_target,img_ori=img$ori_img,smth_win=0.09,
                             row_num=5,col_num=1,cut_coord=cut_coord)
}
\seealso{
\code{\link{colony_auto_crop}}   \code{\link[EBImage]{display}}
}
