% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_array.R
\name{feature_array}
\alias{feature_array}
\title{Compute the image features for the colonies after edge detection}
\usage{
feature_array(obj, img_intensity, type, sample_name, row_num, col_num)
}
\arguments{
\item{obj}{An array containing labelled objects. Labelled objects are pixel sets with the same unique integer value.}

\item{img_intensity}{A matrix or a list of matrices, containing the intensity values of the colony images.}

\item{type}{A character vector containing the function names to be called to compute features with reference intensities.
The choices are \code{"basic"},\code{"moment"} \code{"shape"} and \code{"haralick"}.}

\item{sample_name}{A character vector containing the names of colonies(mutants),the order is column-based(top-down
for every column of mutant array).}

\item{row_num}{An integer specifies the number of the colony need analysis in horizontal direction.}

\item{col_num}{An integer specifies the number of the colony need analysis in vertical direction(for
the condition of only a few colonies,fix it to 1).}
}
\value{
A matrix of n coloies(mutants) times p features, where p depends on the options of feature types given
        to the function.Returns NULL if no object is present.
}
\description{
Computes morphological and texture features from the segmented colonies.
}
