% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_normalize.R
\name{img_normalize}
\alias{img_normalize}
\title{Normalize an original phenotypic image into the gray mode and specific size.}
\usage{
img_normalize(dir = "the directory storing the raw phenotypic image(s)")
}
\arguments{
\item{dir}{a string that gives the directory storing the raw phenotypic image(s)}
}
\value{
a list contains two 'Image' objects or array stores original and normalized
        phenotypic image,respectively
}
\description{
The collected experimental images are always in color mode with various size, which should be
normalzied into gray mode and unified resonable size before analysis.
}
\examples{
img<-img_normalize("directory to/expr_swarming.jpg")
}
