% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_adj_edge.R
\name{manual_adj_edge}
\alias{manual_adj_edge}
\title{Manual adjustment for edge detection of some microorganism colonies}
\usage{
manual_adj_edge(expr_mode = "Congo_red", row_id, col_id, intensity,
  labels, Canny_sigma = 7, low_thr = 6, high_thr = 20,
  mor_size = 9, high_connectivity = TRUE)
}
\arguments{
\item{expr_mode}{A string indicates the phenotypic experiment type, as different experiments require differnt
processing steps}

\item{row_id}{An integer specifies the position of the colony need adjustment in horizontal direction.}

\item{col_id}{An integer specifies the position of the colony need adjustment in vertical direction(for
the condition of only a few colonies,fix it to 1).}

\item{intensity}{A matrix specifies the intensity value of a colony patch,whose value is normalized into [0,1]}

\item{labels}{An array containing labelled objects,which are pixel sets with the same unique integer value to
distinguish colony edge and background.}

\item{Canny_sigma}{A numeric specifies the Gaussian filter variance in Canny detector.}

\item{low_thr}{lower threshold value of the Canny detector.}

\item{high_thr}{upper threshold value of the Canny detector.}

\item{mor_size}{A numeric containing the size of the brush in pixels. This should be an odd number; even numbers
are rounded to the next odd one, i.e., size = 4 has the same effect as size = 5.}

\item{high_connectivity}{A logic 4(false)- or 8(true)-connectivity in 2d case, and between 6(false)- or 26(true)-connectivity in 3d case.
Default is TRUE.}
}
\value{
A list contains 1) An array containing labelled objects. Labelled objects are pixel sets with the same unique integer value.
         2) An array containing labelled objects after the ROI within the edge is filled with the same integer value.
         3) An array, containing the painted version of a certain colony.
}
\description{
As the unified parameters \code{Canny_sigma},\code{low_thr},\code{high_thr} and \code{mor_size}may not fit for some
colonies, \code{manual_adj_edge} is designed for adjusting the edge detection manually for the colonies in need.
}
\examples{
rslt<-manual_adj_edge(expr_mode="Congo_red",i,j,Canny_sigma=2,
                                low_thr=12,high_thr=25,mor_size=13)

}
