% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_act_regionfill.R
\name{pro_act_regionfill}
\alias{pro_act_regionfill}
\title{Fill the ROI of proteolytic activity assay using the same integer value to specify the colony edge}
\usage{
pro_act_regionfill(labels0)
}
\arguments{
\item{labels0}{An array containing labelled objects,which are pixel sets with the same unique integer value to
distinguish colony edge and background}

\item{row_id}{An integer specifies the position of the colony need adjustment in horizontal direction}

\item{col_id}{An integer specifies the position of the colony need adjustment in vertical direction(for
the condition of only a few colonies,fix it to 1).}
}
\value{
An array containing labelled objects after the ROI within the edge is filled with the same integer value.
}
\description{
The ROI of proteolytic activity assay is near a ring, if we want to fill in the area within the ROI, a specific function
\code{pro_act_regionfill} is designed here to accomplish the task.
}
\examples{

}
